refactor(kg): extract entity_extraction + reweave phases to kg/ module
