+++ crates/origin-core/src/kg/mod.rs
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: Apache-2.0
+//! Knowledge graph phase modules: per-phase logic for entity extraction and entity link reweaving.
+
+pub mod entity_extraction;
+pub mod reweave;
